@@ -13,8 +13,8 @@
 		"node-sass": "^4.11.0",
 		"pimg": "^1.1.3",
 		"pm2": "^3.2.9",
-		"react": "^16.7.0",
-		"react-dom": "^16.7.0",
+		"react": "^16.8.3",
+		"react-dom": "^16.8.3",
 		"react-scripts": "2.1.3",
 		"serve-favicon": "^2.5.0"
 	},
