@@ -1,17 +1,13 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import './OpenSource.scss';
 import Sidebar from '../Sidebar/Sidebar';
 
-export default class OpenSource extends React.Component {
-	state = {
-		loading: true,
-		data: {}
-	};
+const OpenSource = () => {
+	const [data, setData] = useState({});
+	const [loading, setLoading] = useState(true);
 
-	componentDidMount() {
-		this.setState({ loading: true });
-
-		setTimeout(async () => {
+	useEffect(() => {
+		(async () => {
 			let response = await fetch(
 				'https://api.github.com/users/ooade/repos?per_page=100'
 			);
@@ -24,35 +20,35 @@ export default class OpenSource extends React.Component {
 				stargazers_count,
 				description
 			}));
-			this.setState({ data, loading: false });
-		}, 1500);
-	}
 
-	render() {
-		const { data, loading } = this.state;
+			setData(data);
+			setLoading(false);
+		})();
+	});
 
-		return (
-			<div className="open-source page" id="open-source">
-				<Sidebar page="Open Source Projects" />
-				<main>
-					{!loading ? (
-						data.map(({ id, name, stargazers_count, description }) => (
-							<div className="open-source__box" key={id}>
-								<h3> {name} </h3>
-								<p> ~ {stargazers_count} Stars </p>
-								<p> {description.replace(/:.*:/, '').trim()} </p>
-							</div>
-						))
-					) : (
-						<div>Loading...</div>
-					)}
-					{!loading && (
-						<div className="open-source--link__box">
-							<a href="https://github.com/ooade?tab=repositories">View All.</a>
+	return (
+		<div className="open-source page" id="open-source">
+			<Sidebar page="Open Source Projects" />
+			<main>
+				{!loading ? (
+					data.map(({ id, name, stargazers_count, description }) => (
+						<div className="open-source__box" key={id}>
+							<h3> {name} </h3>
+							<p> ~ {stargazers_count} Stars </p>
+							<p> {description.replace(/:.*:/, '').trim()} </p>
 						</div>
-					)}
-				</main>
-			</div>
-		);
-	}
-}
+					))
+				) : (
+					<div>Loading...</div>
+				)}
+				{!loading && (
+					<div className="open-source--link__box">
+						<a href="https://github.com/ooade?tab=repositories">View All.</a>
+					</div>
+				)}
+			</main>
+		</div>
+	);
+};
+
+export default OpenSource;
